+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -> SDK 2.0.0-beta.2
 
 -   upgrade the runtime to version 2.0.0-beta.15
+-   enable GET of Files via truncated reference
 
 ## 3.0.0-beta.2